feat: combine role_admin into admin cog for proper subgroup handling
- Merge role_admin.py functionality into admin.py as single AdminCog
- Define admin group and role subgroup as class attributes (matches CTF pattern)
- Delete role_admin.py and its test file
- Revert special loading order in __main__.py (no longer needed)

This fixes the py-cord issue where subgroups defined at module level don't
get their cog attribute set properly. The py-cord examples and working patterns
(CtfCog) define command groups as class attributes inside the Cog.
